+++ .phan/plugins/DuplicateArrayKeyPlugin.php
@@ -58,7 +58,8 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
         }
 
         $case_constant_set = [];
-        foreach ($children as $case_node) {
+        $values_to_check = [];
+        foreach ($children as $i => $case_node) {
             $case_cond = $case_node->children['cond'];
             if ($case_cond === null) {
                 continue;  // This is `default:`. php --syntax-check already checks for duplicates.
@@ -73,10 +74,15 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
             }
             if (is_string($case_cond)) {
                 $cond_key = "s$case_cond";
+                $values_to_check[$i] = $case_cond;
             } elseif (is_int($case_cond)) {
                 $cond_key = $case_cond;
+                $values_to_check[$i] = $case_cond;
             } else {
                 $cond_key = json_encode($case_cond);
+                if (is_scalar($case_cond)) {
+                    $values_to_check[$i] = $case_cond;
+                }
             }
             if (isset($case_constant_set[$cond_key])) {
                 $normalized_case_cond = self::normalizeSwitchKey($case_cond);
@@ -90,8 +96,83 @@ class DuplicateArrayKeyVisitor extends PluginAwarePostAnalysisVisitor
                     Issue::REMEDIATION_A,
                     15071
                 );
+                // Add a fake value to indicate loose equality checks are redundant
+                $values_to_check[-1] = true;
             }
             $case_constant_set[$cond_key] = $case_node->lineno;
+        }
+        if (!isset($values_to_check[-1]) && count($values_to_check) > 1 && !self::areAllSwitchCasesTheSameType($values_to_check)) {
+            // @phan-suppress-next-line PhanPartialTypeMismatchArgument array keys are integers for switch
+            $this->extendedLooseEqualityCheck($values_to_check, $children);
+        }
+    }
+
+    /**
+     * @param array<int,mixed> $values_to_check scalar constant values of case statements
+     */
+    private static function areAllSwitchCasesTheSameType(array $values_to_check) : bool {
+        $categories = 0;
+        foreach ($values_to_check as $value) {
+            if (is_int($value)) {
+                $categories |= 1;
+                if ($categories !== 1) {
+                    return false;
+                }
+            } elseif (is_string($value)) {
+                if (is_numeric($value)) {
+                    // This includes float-like strings such as `"1e0"`, which adds ambiguity ("1e0" == "1")
+                    return false;
+                }
+                $categories |= 2;
+                if ($categories !== 2) {
+                    return false;
+                }
+            } else {
+                return false;
+            }
+        }
+        return true;
+    }
+
+    /**
+     * Perform a heuristic check if any element is `==` a previous element.
+     *
+     * This is intended to perform well for large arrays.
+     *
+     * TODO: Do a better job for small arrays.
+     *
+     * @return void
+     * @param array<int,mixed> $values_to_check
+     * @param array<int,mixed> $children an array of scalars
+     */
+    private function extendedLooseEqualityCheck(array $values_to_check, $children) {
+        $numeric_set = [];
+        $fuzzy_numeric_set = [];
+        foreach ($values_to_check as $i => $value) {
+            if (is_numeric($value)) {
+                // For `"1"`, search for `"1foo"`, `"1bar"`, etc.
+                $value = is_float($value) ? $value : filter_var($value, FILTER_VALIDATE_FLOAT);
+                $old_index = $numeric_set[$value] ?? $fuzzy_numeric_set[$value] ?? null;
+                $numeric_set[$value] = $i;
+            } else {
+                $value = (float)$value;
+                // For `"1foo"`, search for `1` but not `"1bar"`
+                $old_index = $numeric_set[$value] ?? null;
+                // @phan-suppress-next-line PhanTypeMismatchDimAssignment
+                $fuzzy_numeric_set[$value] = $i;
+            }
+            if ($old_index !== null) {
+                $this->emitPluginIssue(
+                    $this->code_base,
+                    clone($this->context)->withLineNumberStart($children[$i]->lineno),
+                    'PhanPluginDuplicateSwitchCaseLooseEquality',
+                    "Switch case({STRING_LITERAL}) is loosely equivalent (==) to an earlier case ({STRING_LITERAL}) in switch statement - the earlier entry may be chosen instead.",
+                    [self::normalizeSwitchKey($values_to_check[$i]), self::normalizeSwitchKey($values_to_check[$old_index])],
+                    Issue::SEVERITY_NORMAL,
+                    Issue::REMEDIATION_A,
+                    15072
+                );
+            }
         }
     }